api: use floats for currency units

+++ network/api.go
@@ -5,22 +5,22 @@ import (
 )
 
 type Block struct {
-	Block       uint32 `json:"block"`
-	EncPubkey   string `json:"enc_pubkey"`
-	Fee         uint64 `json:"fee"`
-	Hashratekhs uint64 `json:"hashratekhs"`
-	Maturation  uint32 `json:"maturation"`
-	Nonce       uint32 `json:"nonce"`
-	Operations  uint64 `json:"operations"`
-	Oph         string `json:"oph"`
-	Payload     string `json:"payload"`
-	Pow         string `json:"pow"`
-	Reward      uint64 `json:"reward"`
-	Sbh         string `json:"sbh"`
-	Target      uint32 `json:"target"`
-	Timestamp   uint32 `json:"timestamp"`
-	Ver         uint16 `json:"ver"`
-	VerA        uint16 `json:"ver_a"`
+	Block       uint32  `json:"block"`
+	EncPubkey   string  `json:"enc_pubkey"`
+	Fee         float64 `json:"fee"`
+	Hashratekhs uint64  `json:"hashratekhs"`
+	Maturation  uint32  `json:"maturation"`
+	Nonce       uint32  `json:"nonce"`
+	Operations  uint64  `json:"operations"`
+	Oph         string  `json:"oph"`
+	Payload     string  `json:"payload"`
+	Pow         string  `json:"pow"`
+	Reward      float64 `json:"reward"`
+	Sbh         string  `json:"sbh"`
+	Target      uint32  `json:"target"`
+	Timestamp   uint32  `json:"timestamp"`
+	Ver         uint16  `json:"ver"`
+	VerA        uint16  `json:"ver_a"`
 }
 
 type Account struct {
@@ -32,18 +32,18 @@ type Account struct {
 }
 
 type Operation struct {
-	Account        uint32 `json:"account"`
-	Amount         uint64 `json:"amount"`
-	Block          uint32 `json:"block"`
-	Dest_account   uint32 `json:"dest_account"`
-	Fee            uint64 `json:"fee"`
-	Opblock        uint32 `json:"opblock"`
-	Ophash         string `json:"ophash"`
-	Optxt          string `json:"optxt"`
-	Optype         uint8  `json:"optype"`
-	Payload        string `json:"payload"`
-	Sender_account uint32 `json:"sender_account"`
-	Time           uint32 `json:"time"`
+	Account        uint32  `json:"account"`
+	Amount         float64 `json:"amount"`
+	Block          uint32  `json:"block"`
+	Dest_account   uint32  `json:"dest_account"`
+	Fee            float64 `json:"fee"`
+	Opblock        uint32  `json:"opblock"`
+	Ophash         string  `json:"ophash"`
+	Optxt          string  `json:"optxt"`
+	Optype         uint8   `json:"optype"`
+	Payload        string  `json:"payload"`
+	Sender_account uint32  `json:"sender_account"`
+	Time           uint32  `json:"time"`
 }
 
 type API interface {
